+++ NEWS.md
@@ -1,5 +1,16 @@
 # antaresEditObject 0.5.2
 
+### Breaking changes (Antares v8.6, cf. Antares v8.6 changelog) :
+
+* `createStudy()` integrate ST-storages.
+* `createArea()` / `removeArea()` integrate ST-storages.
+
+
+NEW FEATURES (Antares v8.6) :
+
+* New function `activateST()` to use new ST features in an Antares study.
+
+
 ### Breaking changes
 
 * `deleteStudy()` no longer requires user confirmation